requirements.txt
djangorestframework
beautifulsoup4==4.4.1
Brlapi==0.6.4
chardet==2.3.0
command-not-found==0.3
defer==1.0.6
Django==2.0.4
html5lib==0.999
language-selector==0.1
louis==2.6.4
lxml==3.5.0
onboard==1.2.0
pexpect==4.0.1
Pillow==3.1.2
ptyprocess==0.5
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
python-apt===1.1.0.b1-elementary2-r1-ubuntu16.04.1
python-systemd==231
pytz==2018.4
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.10.0
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
xkit==0.0.0
